+++ classes/Controllers/Compatibility.php
@@ -30,6 +30,7 @@ class Compatibility extends Controller {
 			'Compatibility\Backcompat\Filters'     => new \PopupMaker\Controllers\Compatibility\Backcompat\Filters( $this->container ),
 			'Compatibility\SEO\Yoast'              => new \PopupMaker\Controllers\Compatibility\SEO\Yoast( $this->container ),
 			'Compatibility\Builder\Divi'           => new \PopupMaker\Controllers\Compatibility\Builder\Divi( $this->container ),
+			'Compatibility\Builder\Elementor'      => new \PopupMaker\Controllers\Compatibility\Builder\Elementor( $this->container ),
 			'Compatibility\Plugin\ACF'             => new \PopupMaker\Controllers\Compatibility\Plugin\ACF( $this->container ),
 			'Compatibility\Plugin\PostDuplication' => new \PopupMaker\Controllers\Compatibility\Plugin\PostDuplication( $this->container ),
 		] );